Renamed electrical sensor files as they were the same as the cluster files
