+++ examples/energy-management/electrical-sensor/include/PowerTopologyDelegateImpl.h
